Move all notification/alerts to subdir
