+++ www/src/js/views/components/notfications/Announcements.scss
@@ -1,4 +1,4 @@
-@import '~styles/utils/modules-entry';
+@import '../../../../styles/utils/modules-entry';
 
 .announcement {
   display: flex;
