+++ www/src/js/views/timetable/TimetableContent.jsx
@@ -35,9 +35,9 @@ import {
 import config from 'config';
 import ModulesSelectContainer from 'views/timetable/ModulesSelectContainer';
 import CorsNotification from 'views/components/cors-info/CorsNotification';
-import Announcements from 'views/components/Announcements';
+import Announcements from 'views/components/notfications/Announcements';
 import Title from 'views/components/Title';
-import RefreshPrompt from 'views/components/RefreshPrompt';
+import RefreshPrompt from 'views/components/notfications/RefreshPrompt';
 import NoLessonWarning from 'views/timetable/NoLessonWarning';
 import Timetable from './Timetable';
 import TimetableActions from './TimetableActions';
